requirements.txt
flask
flask_login
flask_restplus
flask_sqlalchemy
pymysql
flask_marshmallow
marshmallow_sqlalchemy
flask_cors
flask_compress
flask_jwt_extended
flask_migrate
pandas
